requirements.txt
anthropic>=0.25.0
PyQt6>=6.6.0
python-dotenv>=1.0.0
